+++ EsKitsLibrary/src/main/java/io/github/xuyao5/dkl/eskits/service/MySQL2EsService.java
@@ -7,7 +7,7 @@ import com.lmax.disruptor.EventFactory;
 import io.github.xuyao5.dkl.eskits.context.AbstractExecutor;
 import io.github.xuyao5.dkl.eskits.context.DisruptorBoost;
 import io.github.xuyao5.dkl.eskits.context.annotation.TableField;
-import io.github.xuyao5.dkl.eskits.context.disruptor.TwoArgEvent;
+import io.github.xuyao5.dkl.eskits.context.event.TwoArgEvent;
 import io.github.xuyao5.dkl.eskits.dao.InformationSchemaDao;
 import io.github.xuyao5.dkl.eskits.schema.base.BaseDocument;
 import io.github.xuyao5.dkl.eskits.schema.mysql.Columns;
